Refactor the failing encoding tests to avoid field ordering issues (fixes #2)
The failing tests raised in
https://bitbucket.org/ssaasen/haskell-jwt/issue/2/tests-failing were
consistently failing on Linux using the JWT encoding produced on my Mac due
differences in the order of the encoded name/value pairs.

Fields in JSON objects are unordered name/value pairs
(http://www.ietf.org/rfc/rfc4627.txt). Therefore the base64 encoded JSON
representation is not necessarily stable as the header and potentially the
claims contain more than one name/value pair.

To sidestep the issue the tests are now not relying on the String
representation of the JWTs whilst still ensuring the intended behaviour.