Audience claim more closely adheres to JWT spec
The Audience claim may be a list of `StringOrUri`. Therefore,
this pull request changes the representation of `aud` to a
`Maybe (NonEmpty StringOrUri)`. This nicely covers three cases:

1. No `aud` claim is present
2. An `aud` claim that represents a single audience.
3. An `aud` claim that represents more than one audience.

The first case is covered by the presence of `Maybe`. The second,
case makes it simple for someone to grab a single audience if
they anticipate there will only be one audience in the `aud` claim.
And the third case covers the situation where there may be many
audiences in the `aud` claim.

This pull request adds a dependencies on `semigroups` (for the
`NonEmpty data type) and `vector` (for the `JSON` processing).

The JSON instances pretty much copied the the List instances in
`aeson`.