Introduce 'doctests' flag and make the doctests disabled by default
Due to the problem encountered in #2
(https://bitbucket.org/ssaasen/haskell-jwt/issue/2/tests-failing) the doctests
are failing on Linux. As the doctests are mainly showcasing usage examples they
should not be rewritten in the way #2 was resolved. Instead 'cabal test' now
does not run the doctests unless specifically told to do so using the new flag.